pull way types from parse.hpp
