+++ include/elastic/archive_type.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include "types/message.hpp"
+#include "types/sequence.hpp"
+#include "types/types.hpp"
