+++ include/elastic/types/message.hpp
@@ -1,0 +1,47 @@
+#pragma once
+#include "../reflect.hpp"
+#include "../archive_exception.hpp"
+
+namespace
+{
+	template <typename _Ty, typename _Func, std::size_t... I>
+	constexpr void for_each(_Ty&& val, _Func&& func, std::index_sequence<I...>)
+	{
+		return (std::forward<_Func>(func)(elastic::get<I>(std::forward<_Ty>(val))), ...);
+	}
+
+	template <typename _Ty, typename _Func, std::size_t N = elastic::tuple_size_v<_Ty>,
+			  typename Indices = std::make_index_sequence<N>>
+	constexpr void for_each(_Ty&& val, _Func&& func)
+	{
+		return for_each(std::forward<_Ty>(val), std::forward<_Func>(func), Indices{});
+	}
+}
+
+namespace elastic
+{
+	template <typename _Ty, typename _Archive>
+	struct message
+	{
+		static void deserialize(_Archive& ar, _Ty& t)
+		{
+			constexpr auto N = elastic::tuple_size_v<_Ty>;
+
+			using Indices = std::make_index_sequence<N>;
+
+			for_each(t,
+					 [&](auto&& v)
+					 {
+						 ar >> v;
+
+						 if (ar.interrupt())
+							 throw(archive_exception::exception_number::output_stream_error, "make element error!");
+					 });
+		}
+
+		static void serialize(_Ty&& value, _Archive& ar)
+		{
+			for_each(std::forward<_Ty>(value), [&](auto&& v) { ar << v; });
+		}
+	};
+}
