+++ test/test.vcxproj
@@ -137,6 +137,7 @@
     <ClInclude Include="..\include\elastic.hpp" />
     <ClInclude Include="..\include\elastic\access.hpp" />
     <ClInclude Include="..\include\elastic\archive_exception.hpp" />
+    <ClInclude Include="..\include\elastic\archive_type.hpp" />
     <ClInclude Include="..\include\elastic\basic_iprimitive.hpp" />
     <ClInclude Include="..\include\elastic\basic_oprimitive.hpp" />
     <ClInclude Include="..\include\elastic\basic_primitive.hpp" />
@@ -154,10 +155,12 @@
     <ClInclude Include="..\include\elastic\detail\size_t_.hpp" />
     <ClInclude Include="..\include\elastic\detail\type_traits.hpp" />
     <ClInclude Include="..\include\elastic\interface_archive.hpp" />
-    <ClInclude Include="..\include\elastic\parser.hpp" />
     <ClInclude Include="..\include\elastic\reflect.hpp" />
     <ClInclude Include="..\include\elastic\tuple_size.hpp" />
+    <ClInclude Include="..\include\elastic\types\message.hpp" />
+    <ClInclude Include="..\include\elastic\types\sequence.hpp" />
     <ClInclude Include="..\include\elastic\types\types.hpp" />
+    <ClInclude Include="..\include\elastic\types\varint.hpp" />
     <ClInclude Include="attr.h" />
     <ClInclude Include="exception.h" />
     <ClInclude Include="generate.h" />
