+++ test/test.vcxproj.filters
@@ -110,13 +110,22 @@
     <ClInclude Include="..\include\elastic\interface_archive.hpp">
       <Filter>elastic</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\elastic\parser.hpp">
-      <Filter>elastic</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\elastic\archive_exception.hpp">
       <Filter>elastic</Filter>
     </ClInclude>
     <ClInclude Include="..\include\elastic\basic_primitive.hpp">
+      <Filter>elastic</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\elastic\types\varint.hpp">
+      <Filter>elastic\types</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\elastic\types\sequence.hpp">
+      <Filter>elastic\types</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\elastic\types\message.hpp">
+      <Filter>elastic\types</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\elastic\archive_type.hpp">
       <Filter>elastic</Filter>
     </ClInclude>
   </ItemGroup>
